Add preliminary ModelForm support.
Not totally sure that this covers all cases (and it certainly is going
to make the default create form sad, since at that point we don’t know
what type a model is), but it works for an existing model’s change view.

I still don’t have a good solution in terms of flow for adding a new
model (best case is a two step form, which isn’t terrible, and could be
spruced up with some js polling for the form once the user selects a
type), and changing a model’s type (basically the same as the two step
form, but a tad bit trickier in terms of what the user expectations are
for data preservation between types).